feat: Add about page
This commit adds a new file `about.astro` which contains the content for the about page. The page provides information about the software engineer, Duc An, who is based in Vietnam. The content will be updated soon.

This commit enhances the website by adding an about page, improving the user experience and providing more information to visitors.