feat: Update font and image services, improve date formatting, and add about page
This commit updates the font and image services in the `astro.config.mjs` file, replacing the `passthroughImageService` with `sharpImageService` and `squooshImageService`. It also improves the date formatting in the `FormattedDate.astro`, `index.astro`, and `bookmarks/index.astro` components by using the `format` function from the `date-fns` library. Additionally, it adds a new file `about.astro` which contains the content for the about page, providing information about the software engineer, Duc An. These changes enhance the website by improving the visual appeal, user experience, and providing more information to visitors.